Fix HTML select menupopup appearance + scrolling This commit fixes the appearance of the HTML select's menupoup by making certain styles not apply to it.
It also does the harder task of fixing the ability to scroll this menu
in Firefox 115. Previously, to change the tab scrollbox arrow images,
an override of chrome://global/skin/arrowscrollbox.css was used. I
assume this was taken from a newer version of Firefox due to its use of
nesting blocks (despite it being compiled from SCSS in Namoroka).
Regardless, this override was the culprit to HTML select popups not
being scrollable. I fixed this by ditching the override and instead
creating a new separate file (tabscrollbox.uc.scss) which is manually
added to the tab scrollbox's shadow root via JavaScript.